fix(nxp-mmc): correct the usage of BIT and GENMASK
The uSDHC register definitions use BIT and GENMASK macros, which are
implicitly expanded to their 64-bit versions. This is incorrect from a
hardware perspective, as the registers are 32-bit wide. Therefore,
BIT_32 and GENMASK_32 macros are used instead. Similarly, the register
offset must use an unsigned int offset instead of signed numbers.

Change-Id: I3a3379ba9fa538226f958b68aac702752ea9a62a
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>