refactor(memmap): factor memory regions into their own structure
This is a non-functional change to clean up how we represent memory
regions in the memory map tool.

This introduces the `Region` data-class, which does little more than
store the region data exposed by the linker scripts, and provides some
additional helper methods which re-produce what was already being done
by both the ELF parser and the map file parser.

Change-Id: Idf45b71bd460f7312498b97b9e4743f4dc482411
Signed-off-by: Chris Kay <chris.kay@arm.com>