feat(mt8196): show ERROR log if need
There are two purposes for hardware semaphores. The first one is for
the SMMU to check the NPU power status, and the second one is for NPU
power control.

In the case of the SMMU, if the hardware semaphore cannot be locked
immediately, it means the NPU is powered off, and simply returning
-EBUSY is sufficient.

Hence, there is no need to show any ERROR message for the SMMU case.
(retry_times == HW_SEM_NO_WAIT)

Change-Id: I1a0b6c16e297c7564518883863ebc67e38b6561a
Signed-off-by: Chungying Lu <chungying.lu@mediatek.com>