refactor(memmap): represent image symbols with key-value map
The symbols defined by an image are currently returned as a list of
tuple pairs, which is a fairly unintuitive way of representing what is
essentially a map.

This commit changes the return type to the appropriate map type, and
moves the heavy-lifting required to turn it into a list of tuples into
the printer, which is the only place where this representation is
necessary.

Change-Id: I87d047608b390e1e0cdf93eb6af5aeb103b1d69e
Signed-off-by: Chris Kay <chris.kay@arm.com>