fix(sptool): handle load-address-relative-offset property
If the SP manifest contains load-address-relative-offset then the script
failed to generate the sp_list_fragment.dts file because that line was
also added to the load_address_lines list.
This commit fixes this issue: a line is only added to the list if the
property name is exactly "load-address", without a following '-'
character.

Change-Id: I1fd2006d37d9cc1864af31576e60565a7bce6023
Signed-off-by: Bence Balogh <bence.balogh@arm.com>