fix(st-uart): aarch64: unwait for UART ISR register TC bit in putc function
The loop over the UART_ISR_TC flag was needed before the UART FIFO was
enabled. This allowed to make sure each character written in TDR was
outputed.
This behavior is no more needed. Once a character is in the FIFO the
UART will empty it when it is clocked.

Change-Id: I914c7f75a451bedbcc9287d8ed9178db47b4eab4
Signed-off-by: Clément Le Goffic <legoffic.clement@gmail.com>