fix(lx2160a): put cert_create_tbbr.mk in the standard location
The cert_create_tbbr.mk file is generally looked for in the platform
directory. The lx2160a breaks from this convention and stores the file
elsewhere, requiring a cludgy `cp` mid way through the make process. The
result is an untracked file when building the platform, which is
undesirable.

Fix this by putting the cert_create_tbbr.mk (and associated source
files) in the plat/ directory, just like we do for juno and stm32mp1.

Change-Id: Ifa8bb07dba4f55da4f0e230efa7beccf30bd1de7
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>