fix(drtm): add event name data for separator event entries
Update the DRTM measurement logging to include `drtm_event_arm_sep_data`
as the event name for separator events recorded in PCR-17 and PCR-18.

Previously, the event name parameter was set to `NULL`, resulting in
incomplete event records. This fix ensures that separator events are
correctly labeled in the event log with meaningful context and it
aligns with DRTM specification v1.1.

Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>
Change-Id: I2803d0f2e113352116d7894c2a89dfc48ed0fd85