fix(drtm): set correct max region mask for DMA protection to 0xFFFF
Update `ARM_DRTM_DMA_PROT_FEATURES_MAX_REGIONS_MASK` from `0xF` to
`0xFFFF` to correctly reflect the maximum number of supported DMA
protection regions, as per the expected DRTM capability encoding.

The previous 4-bit mask limited the range to 16 regions, which is
insufficient and incorrect based on the spec's intention. This fix
extends the bitmask to 16 bits, allowing up to 65,535 regions if
supported in future implementations.

Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>
Change-Id: I2531706f4b75125b510413870dfd4b1804e5adb6