fix(ti): remove validate_power_state definition
validate_power_state should not be defined since CPU suspend is not yet
supported.

Defining the validate_power_state function without any meaningful logic
caused a bug where a crash would happen if a device tried to enter a power
state that was not yet supported by TFA. This is because
validate_power_state would return a pass when the state was not properly
checked.

Removing the validate_power_state definition disables the
PSCI_CPU_SUSPEND_AARCH64 capability which will prevent the HLOS from
trying to suspend the CPU all together. This will fix the aforementioned
bug from occurring.

Fixes: 2e9c9e829964 ("ti: k3: common: Add PSCI stubs")
Change-Id: I3bbbd228fbddea64e72f0cf50afc6b25fb6d317c
Signed-off-by: Kendall Willis <k-willis@ti.com>