fix(versal2): validate Node ID in PM callback events
The PM_NOTIFY_CB is used to notify both events and errors from the PLM
firmware. The values of events and errors can overlap, they can be
distinguished using the Node ID, which is included in the callback payload.

Currently, when an EVENT_CPU_PWRDWN notification is received, TF-A powers
down cores without validating the Node ID. This leads to incorrect
behavior, as TFA powers down cores even when an error occurs that shares
the same event value.

Add a Node ID check to differentiate between events and errors to fix
this issue.

Change-Id: I65d69731b692928597e47678c684aea2b90b5e6d
Signed-off-by: Naman Trivedi <naman.trivedimanojbhai@amd.com>