build(commitlint): treat wrong scopes as error
The commitlint checker treats scopes that are not listed in the
changelog.yaml file as warnings. This is then usually overlooked, as
commitlint checker passes with warnings in CI and some patches with
wrong scope can be merged. The maintainers doing the release note will
then have to deal with those patches manually.

Signed-off-by: Yann Gautier <yann.gautier@st.com>
Change-Id: Ica5f581ce7b8c1adbe35f5ab3e6870d9a07807fd