PENDING: fix(ti): core ON/OFF while suspend resume
Key observation: using the scmi_handler code for turn ON/OFF seems to be
causing issues, possibly because the core continues to poll on it's own
MDSTAT which will obviously timeout and not let the core shut down
cleanly. This causes issues when we try to turn on the core on resume,
it's possibly left in an unclean state and hence refuses to turn on.

Fixes: ef67277f37c4 ("FROMLIST: feat(ti): PSCI: Add support for dual core CPU in AM62L")
Change-Id: I131b83635ea5b6080870ae1b18e5665d1e6fdd00
Signed-off-by: Dhruva Gole <d-gole@ti.com>