fix(fwu): fwu NV ctr upgraded on trial run
The NV ctr value should not upgraded on trial run.
The NV ctr value upgrade is done in BL1 while the detection
of trial run happens in BL2, so the value is always upgraded.
Fix the problem by setting the upgrade of NV ctr value in BL2
if the NV ctr is shared among components.

Change-Id: Id681fce0482e3000eaef4f4a8f7d8c1023ccaf1a
Signed-off-by: Xialin Liu <xialin.liu@arm.com>